requirements.txt
pandas==2.2.0
numpy==1.26.0
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
nebula3-python==3.5.0
google-generativeai==0.3.0
requests==2.31.0